Refactor object writing responsiblities to ObjectDatabase
The DatabaseInserter API permits ObjectDatabase implementations to
control their own object insertion behavior, rather than forcing
it to always be a new loose file created in the local filesystem.

This change starts to open the door to non-file object storage,
such as a trivial in-memory HashMap for unit testing, or a more
complex system built on top of a distributed hash table.

To help existing application code port to a newer interface we
are keeping ObjectWriter as a delegation wrapper to the new API.
Each ObjectWriter instances holds a reference to a DatabaseInserter
for the Repository's top-level ObjectDatabase.

Signed-off-by: Shawn O. Pearce <spearce@spearce.org>